// src/main/webapp/src/app/rest/model/gitHubPrCriteriaDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GitHubPrCriteriaDTO { 
    fromYear?: number;
    toYear?: number;
    usernamePattern?: string;
    fromPullRequestNumber?: number;
    toPullRequestNumber?: number;
    pullRequestNumberPattern?: string;
    titleContains?: string;
    bodyContains?: string;
    authorContains?: string;
    labelContains?: string;
    baseRefContains?: string;
    excludedStates?: string;
    draftAvailability?: string;
    mergedAvailability?: string;
    mergeableAvailability?: string;
    mergeableStatePattern?: string;
    analysisSummaryContains?: string;
    analysisUserExtraPromptsContains?: string;
    analysisSummaryUpdatedFrom?: string;
    analysisSummaryUpdatedTo?: string;
    analysisSummaryMinTokensK?: number;
    analysisSummaryMaxTokensK?: number;
    analysisAvailability?: string;
    developmentWorkDescribedContains?: string;
    developmentWorkUserExtraPromptsContains?: string;
    developmentWorkUpdatedFrom?: string;
    developmentWorkUpdatedTo?: string;
    developmentWorkMinTokensK?: number;
    developmentWorkMaxTokensK?: number;
    developmentWorkAvailability?: string;
    personalInterrestCommentContains?: string;
    personalInterrestMinPriority?: number;
    personalInterrestMaxPriority?: number;
    personalInterrestAvailability?: string;
}